Refactor contact_email validation as a mixin
