+++ app/controllers/concerns/can_resend_email.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+
+module CanResendEmail
+  extend ActiveSupport::Concern
+
+  private
+
+  def validate_contact_email(registration)
+    raise Exceptions::MissingContactEmailError, registration.reg_identifier unless registration.contact_email.present?
+
+    assisted_digital_match = registration.contact_email == WasteCarriersEngine.configuration.assisted_digital_email
+
+    raise Exceptions::AssistedDigitalContactEmailError, registration.reg_identifier if assisted_digital_match
+  end
+
+  # If the registration has a missing contact_email we know
+  # RenewalReminderMailer will throw a MissingContactEmailError. There is not
+  # really anything we can do and the reason for the missing field is legacy
+  # issues. So we don't worry about Airbrake or adding anything to the log.
+  def handle_missing_contact_email(scope)
+    message = I18n.t("#{scope}.messages.missing.heading")
+    description = I18n.t("#{scope}.messages.missing.details")
+
+    flash_error(message, description)
+  end
+
+  # If the registration has a contact_email that matches the configured assisted
+  # digital one we know RenewalReminderMailer will throw a
+  # AssistedDigitalContactEmailError. This isn't something we want to record. So
+  # we don't worry about Airbrake or adding anything to the log.
+  def handle_assisted_digital_contact_email(scope)
+    message = I18n.t("#{scope}.messages.assisted.heading")
+    description = I18n.t(
+      "#{scope}.messages.assisted.details",
+      email: WasteCarriersEngine.configuration.assisted_digital_email
+    )
+
+    flash_error(message, description)
+  end
+
+  # Here we do notify Airbrake and log an error. If we get here it's for an
+  # unexpected reason hence we want to know about it.
+  def handle_resend_errored(error, scope, applies_to)
+    Airbrake.notify error, registration: registration.reg_identifier
+    Rails.logger.error "Error #{applies_to} for registration #{registration.reg_identifier}"
+
+    message = I18n.t("#{scope}.messages.error.heading", email: registration.contact_email)
+    description = I18n.t("#{scope}.messages.error.details")
+
+    flash_error(message, description)
+  end
+end
